requirements.txt
google-adk==1.18.0
python-dotenv==1.2.1
scholarly==1.7.11
